+++ packages/common/src/types/compression.ts
@@ -1,5 +1,7 @@
-export type CompressedPayload = `${"0" | "1"}${"0" | "1"}${"0" | "1"}${
+export type CompressedDiagram = `${"0" | "1"}${"0" | "1"}${"0" | "1"}${
   | "0"
-  | "1"}${"0" | "1"}${"0" | "1"}${string}`;
+  | "1"}${"0" | "1"}${"0" | "1"}`;
+
+export type CompressedPayload = `${CompressedDiagram}${string}`;
 
 export type Dictionary = { [x: string]: string };
