+++ packages/common/src/utils/index.spec.ts
@@ -42,15 +42,19 @@ describe("test utils / isExpired", function () {
 
 describe("test utils / areAllSegmentsPresent", function () {
   it("should return true", async function () {
-    const completeSegments = (await createMockedFlow(RPC_REQ_LARGE).next())
-    .value.toMessage()
-    .toSegments();
+    const completeSegments = (
+      await createMockedFlow(RPC_REQ_LARGE).next()
+    ).value
+      .toMessage()
+      .toSegments();
     assert(areAllSegmentsPresent(completeSegments));
   });
   it("should return false", async function () {
-    const completeSegments = (await createMockedFlow(RPC_REQ_LARGE).next())
-    .value.toMessage()
-    .toSegments();
+    const completeSegments = (
+      await createMockedFlow(RPC_REQ_LARGE).next()
+    ).value
+      .toMessage()
+      .toSegments();
     const incompleteSegments = completeSegments.slice(1);
     assert(!areAllSegmentsPresent(incompleteSegments));
   });
@@ -58,16 +62,20 @@ describe("test utils / areAllSegmentsPresent", function () {
 
 describe("test utils / areAllSegmentsPresent", function () {
   it("should return true", async function () {
-    const completeSegments = (await createMockedFlow(RPC_REQ_LARGE).next())
-    .value.toMessage()
-    .toSegments();
+    const completeSegments = (
+      await createMockedFlow(RPC_REQ_LARGE).next()
+    ).value
+      .toMessage()
+      .toSegments();
 
     assert(areAllSegmentsPresent(completeSegments));
   });
   it("should return false", async function () {
-    const completeSegments = (await createMockedFlow(RPC_REQ_LARGE).next())
-    .value.toMessage()
-    .toSegments();
+    const completeSegments = (
+      await createMockedFlow(RPC_REQ_LARGE).next()
+    ).value
+      .toMessage()
+      .toSegments();
     const incompleteSegments = completeSegments.slice(1);
 
     assert(!areAllSegmentsPresent(incompleteSegments));
